backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
numpy>=1.26.0,<2.0.0
librosa>=0.10.1,<0.11.0
pydub>=0.25.1,<0.26.0
soundfile>=0.12.1,<0.13.0
scipy>=1.11.0,<2.0.0
python-multipart>=0.0.6,<0.0.8
openai-whisper>=20231117
jiwer>=3.0.3,<4.0.0
